+++ test/envtest/setup.go
@@ -5,13 +5,14 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
+	"runtime"
 	"sync"
 	"syscall"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/require"
-	"k8s.io/apimachinery/pkg/runtime"
+	k8sruntime "k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
@@ -20,12 +21,31 @@ import (
 	"github.com/kong/kubernetes-ingress-controller/v2/test/consts"
 )
 
+type Options struct {
+	InstallGatewayCRDs bool
+	InstanllKongCRDs   bool
+}
+
+var DefaultEnvTestOpts = Options{
+	InstallGatewayCRDs: true,
+	InstanllKongCRDs:   false,
+}
+
+type OptionModifier func(Options) Options
+
+func WithInstallKongCRDs(install bool) OptionModifier {
+	return func(opts Options) Options {
+		opts.InstanllKongCRDs = install
+		return opts
+	}
+}
+
 // Setup sets up the envtest environment which will be stopped on test cleanup
 // using t.Cleanup().
 //
 // Note: If you want apiserver output on stdout set
 // KUBEBUILDER_ATTACH_CONTROL_PLANE_OUTPUT to true when running tests.
-func Setup(t *testing.T, scheme *runtime.Scheme) *rest.Config {
+func Setup(t *testing.T, scheme *k8sruntime.Scheme, optModifiers ...OptionModifier) *rest.Config {
 	t.Helper()
 
 	testEnv := &envtest.Environment{
@@ -36,13 +56,17 @@ func Setup(t *testing.T, scheme *runtime.Scheme) *rest.Config {
 	cfg, err := testEnv.Start()
 	require.NoError(t, err)
 
-	gatewayCRDPath := filepath.Join(build.Default.GOPATH, "pkg", "mod", "sigs.k8s.io", "gateway-api@"+consts.GatewayAPIVersion, "config", "crd", "experimental")
-	_, err = envtest.InstallCRDs(cfg, envtest.CRDInstallOptions{
-		Scheme:             scheme,
-		Paths:              []string{gatewayCRDPath},
-		ErrorIfPathMissing: true,
-	})
-	require.NoError(t, err, "failed installing Gateway API CRDs")
+	opts := DefaultEnvTestOpts
+	for _, mod := range optModifiers {
+		opts = mod(opts)
+	}
+
+	if opts.InstallGatewayCRDs {
+		installGatewayCRDs(t, scheme, cfg)
+	}
+	if opts.InstanllKongCRDs {
+		installKongCRDs(t, scheme, cfg)
+	}
 
 	wg := sync.WaitGroup{}
 	wg.Add(1)
@@ -81,4 +105,29 @@ func Setup(t *testing.T, scheme *runtime.Scheme) *rest.Config {
 	})
 
 	return cfg
+}
+
+func installGatewayCRDs(t *testing.T, scheme *k8sruntime.Scheme, cfg *rest.Config) {
+	gatewayCRDPath := filepath.Join(build.Default.GOPATH, "pkg", "mod", "sigs.k8s.io", "gateway-api@"+consts.GatewayAPIVersion, "config", "crd", "experimental")
+	_, err := envtest.InstallCRDs(cfg, envtest.CRDInstallOptions{
+		Scheme:             scheme,
+		Paths:              []string{gatewayCRDPath},
+		ErrorIfPathMissing: true,
+	})
+	require.NoError(t, err, "failed installing Gateway API CRDs")
+}
+
+func installKongCRDs(t *testing.T, scheme *k8sruntime.Scheme, cfg *rest.Config) {
+	// extract project root path.
+	_, thisFilePath, _, _ := runtime.Caller(0) //nolint:dogsled
+	projectRoot := filepath.Join(filepath.Dir(thisFilePath), "..", "..")
+	// install Kong CRDs from config/crd/bases.
+	kongCRDPath := filepath.Join(projectRoot, "config", "crd", "bases")
+	t.Logf("install Kong CRDs from manifests in %s", kongCRDPath)
+	_, err := envtest.InstallCRDs(cfg, envtest.CRDInstallOptions{
+		Scheme:             scheme,
+		Paths:              []string{kongCRDPath},
+		ErrorIfPathMissing: true,
+	})
+	require.NoError(t, err)
 }